requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf